Import code thus far...
Parses Maildir and mbox, and produces some estimates of security based on
Received: header classification.

The accuracy of these estimates is uncertain; there are lots of different MTAs
out there, that record their behaviour in various ways, and there may in some
cases be layer IPSEC, MPLS, VPN or other forms of encryption that we can't see.